+++ volmdlr/edges.py
@@ -1248,7 +1248,7 @@ class BSplineCurve2D(BSplineCurve):
         :return: A 2 dimensional point representing the tangent
         :rtype: :class:`volmdlr.Point2D`
         """
-        _, tangent = operations.tangent(self.curve, position,
+        _, tangent = operations.tangent(self.curve, position / self.length(),
                                         normalize=True)
         tangent = volmdlr.Point2D(tangent[0], tangent[1])
         return tangent
@@ -1272,16 +1272,9 @@ class BSplineCurve2D(BSplineCurve):
         normal vector is to be calculated
         :return: The normal vector of the BSplineCurve2D
         """
-        # TODO: The commented version is not exact, and the new one is not efficient
-        # tangent_vector = self.tangent(abscissa)
-        # normal_vector = tangent_vector.normal_vector()
-        # return normal_vector
-
-        wire = self.to_wire(n=20)
-        distances = [primitive.point_distance(self.point_at_abscissa(abscissa))
-                     for primitive in wire.primitives]
-
-        return wire.primitives[distances.index(min(distances))].normal_vector(0.5)
+        tangent_vector = self.tangent(abscissa)
+        normal_vector = tangent_vector.normal_vector()
+        return normal_vector
 
     def unit_normal_vector(self, abscissa: float):
         """